Tidy up & eradicate compilation warnings

+++ app/config/Properties.scala
@@ -15,7 +15,7 @@ trait Properties {
   lazy private val CONFIG: Config = configuration.underlying
 
   // Utils
-  lazy val API_REQUEST_TIMEOUT: Long = CONFIG.getInt("request.timeout")
+  lazy val API_REQUEST_TIMEOUT: Long = CONFIG.getLong("request.timeout")
   lazy val MINIMUM_KEY_LENGTH: Int = CONFIG.getInt("minimum.key.length")
   // SBR CONTROL API
   lazy val CONTROL_EDIT_ENTERPRISE_URL: String = CONFIG.getString("api.sbr.control.edit.enterprise")
